requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
alembic
pydantic
python-dotenv
aiofiles
httpx
pillow
PyMuPDF
pytesseract
openai
anthropic
pandas
openpyxl
requests
urllib3
baidu-aip